feat(hrm): Implement SRS leave calculation logic in timesheets
- Modify `Employee.available_leave_days` to DecimalField.
- Add `generated_leave_days` to `EmployeeMonthlyTimesheet`.
- Implement `calculate_generated_leave` service function based on Contract.
- Update `create_monthly_timesheet_for_employee` to handle Opening, Generated, Carried (w/ Expiration) logic.
- Update `prepare_monthly_timesheets` task to update employee leave balances.
- Update tests to cover new logic and fix regressions.

+++ apps/hrm/models/employee.py
@@ -382,7 +382,9 @@ class Employee(ColoredValueMixin, AutoCodeMixin, BaseModel):
     )
 
     # Available leave days (e.g., remaining annual leave)
-    available_leave_days = models.IntegerField(
+    available_leave_days = models.DecimalField(
+        max_digits=6,
+        decimal_places=2,
         default=0,
         verbose_name="Available leave days",
         help_text="Number of available leave days for the employee",
